替换掉BigDecimal

+++ src/main/java/com/dormitory/dao/ElectricityDAO.java
@@ -1,7 +1,5 @@
 package com.dormitory.dao;
 
-import java.math.BigDecimal;
-
 import org.apache.ibatis.annotations.Result;
 import org.apache.ibatis.annotations.Results;
 import org.apache.ibatis.annotations.Select;
@@ -14,9 +12,9 @@ public interface ElectricityDAO {
 		+" where e.dormitory_id=#{0} ")
 @Results(value = {
         @Result(id = true, property = "dormitoryId", column = "dormitory_id", javaType = Integer.class, jdbcType = JdbcType.INTEGER),
-        @Result(property = "restElectricity", column = "rest_electricity", javaType = BigDecimal.class, jdbcType = JdbcType.DECIMAL),
-        @Result(property = "sumElectricity", column = "sum_electricity", javaType = BigDecimal.class, jdbcType = JdbcType.DECIMAL),
-        @Result(property = "balance", column = "balance", javaType = BigDecimal.class, jdbcType = JdbcType.DECIMAL),
+        @Result(property = "restElectricity", column = "rest_electricity", javaType = Double.class, jdbcType = JdbcType.DECIMAL),
+        @Result(property = "sumElectricity", column = "sum_electricity", javaType = Double.class, jdbcType = JdbcType.DECIMAL),
+        @Result(property = "balance", column = "balance", javaType = Double.class, jdbcType = JdbcType.DECIMAL),
         })
     Electricity getElectricity(Integer dormitoryId);
